Refactor the initialization of analytics config.
Update the analytics prompt.

Update config imports for consistency.
