+++ cmd/gitops/get/config/cmd.go
@@ -5,13 +5,13 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/weaveworks/weave-gitops/cmd/gitops/config"
+	cfg "github.com/weaveworks/weave-gitops/cmd/gitops/config"
 
-	gitopsConfig "github.com/weaveworks/weave-gitops/pkg/config"
+	"github.com/weaveworks/weave-gitops/pkg/config"
 	"github.com/weaveworks/weave-gitops/pkg/logger"
 )
 
-func ConfigCommand(opts *config.Options) *cobra.Command {
+func ConfigCommand(opts *cfg.Options) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "config",
 		Short: "Prints out the CLI configuration for Weave GitOps",
@@ -27,15 +27,15 @@ gitops get config`,
 	return cmd
 }
 
-func getConfigCommandRunE(opts *config.Options) func(*cobra.Command, []string) error {
+func getConfigCommandRunE(opts *cfg.Options) func(*cobra.Command, []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		var err error
 
 		log := logger.NewCLILogger(os.Stdout)
 
-		cfg, err := gitopsConfig.GetConfig(false)
+		cfg, err := config.GetConfig(false)
 		if err != nil {
-			log.Warningf(gitopsConfig.WrongConfigFormatMsg)
+			log.Warningf(config.WrongConfigFormatMsg)
 			return err
 		}
