+++ cmd/gitops/set/config/cmd.go
@@ -10,8 +10,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/weaveworks/weave-gitops/cmd/gitops/cmderrors"
-	"github.com/weaveworks/weave-gitops/cmd/gitops/config"
-	gitopsConfig "github.com/weaveworks/weave-gitops/pkg/config"
+	cfg "github.com/weaveworks/weave-gitops/cmd/gitops/config"
+	"github.com/weaveworks/weave-gitops/pkg/config"
 	"github.com/weaveworks/weave-gitops/pkg/logger"
 )
 
@@ -21,7 +21,7 @@ const (
 
 var analyticsValue bool
 
-func ConfigCommand(opts *config.Options) *cobra.Command {
+func ConfigCommand(opts *cfg.Options) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "config",
 		Short: "Set the CLI configuration for Weave GitOps",
@@ -59,7 +59,7 @@ func setConfigCommandPreRunE(endpoint *string) func(*cobra.Command, []string) er
 	}
 }
 
-func setConfigCommandRunE(opts *config.Options) func(*cobra.Command, []string) error {
+func setConfigCommandRunE(opts *cfg.Options) func(*cobra.Command, []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		var err error
 
@@ -69,7 +69,7 @@ func setConfigCommandRunE(opts *config.Options) func(*cobra.Command, []string) e
 			log.Warningf("This will only turn off analytics for the GitOps CLI. Please refer to the documentation to turn off the analytics in the GitOps Dashboard.")
 		}
 
-		cfg, err := gitopsConfig.GetConfig(true)
+		cfg, err := config.GetConfig(true)
 		if err != nil {
 			return err
 		}
@@ -79,12 +79,12 @@ func setConfigCommandRunE(opts *config.Options) func(*cobra.Command, []string) e
 		if cfg.UserID == "" {
 			seed := time.Now().UnixNano()
 
-			cfg.UserID = gitopsConfig.GenerateUserID(10, seed)
+			cfg.UserID = config.GenerateUserID(10, seed)
 		}
 
 		log.Actionf("Saving GitOps CLI config ...")
 
-		if err = gitopsConfig.SaveConfig(cfg); err != nil {
+		if err = config.SaveConfig(cfg); err != nil {
 			log.Failuref("Error saving GitOps CLI config")
 			return err
 		}
